+++ src/wavm/CMakeLists.txt
@@ -4,7 +4,7 @@ include_directories(
 )
 
 set(HEADERS
-        "${FAASM_INCLUDE_DIR}/wavm/PlatformThreadPool.h"
+        "${FAASM_INCLUDE_DIR}/wavm/OMPThreadPool.h"
         "${FAASM_INCLUDE_DIR}/wavm/WAVMWasmModule.h"
         )
 
@@ -25,7 +25,7 @@ set(LIB_FILES
         mpi.cpp
         network.cpp
         openmp.cpp
-        PlatformThreadPool.cpp
+        OMPThreadPool.cpp
         process.cpp
         rust.cpp
         scheduling.cpp
