+++ src/wavm/openmp/Level.cpp
@@ -35,7 +35,7 @@ namespace wasm {
             int nextWanted = pushedNumThreads > 0 ? pushedNumThreads : wantedNumThreads;
 
             // Returns user preference if set or device's maximum
-            return nextWanted > 0 ? nextWanted : (int) util::getSystemConfig().maxWorkersPerFunction;
+            return nextWanted > 0 ? nextWanted : (int) util::getSystemConfig().ompThreadPoolSize;
         }
 
         void Level::snapshot_parent(message::Message &msg) const {
